# other/openapi.yaml
openapi: '3.0.2'
info:
  title: GainTracker
  version: '1.0'
  description: |-
    Application to record results in your training sessions.
  termsOfService: http://swagger.io/terms/
  contact:
    email: gaintrackerteam@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:5000
...
paths:
...
  /user/progress/measurements/add:
    post:
      tags:
        - user
      summary: Record your bodyparts
      description: Make new measurement record
      operationId: addMeasurements
      requestBody:
        description: Make measurement record
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/measurementRecordCreation'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/measurementsList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /user/progress/measurements:
    get:
      tags:
        - user
      summary: Get user info and body measurements
      description: ''
      operationId: getUserMeasurementsInfo
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userAndMeasurementsInfo' 
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /user/progress/exercises:
    get:
      tags:
        - user
      summary: Get user info and progress in exercises
      description: ''
      operationId: getUserExercisesInfo
      parameters:
        - name: exercise_name
          in: query
          required: false
          schema:
            type: string
            example: 'Push ups'
            default: 'None'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userAndExerciseInfo' 
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /user/progress/muscles:
    get:
      tags:
        - user
      summary: Get user info and progress in muscle groups
      description: ''
      operationId: getUserMusclesInfo
      parameters:
        - name: muscle_name
          in: query
          required: false
          schema:
            type: string
            example: 'Chest'
            default: 'None'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userAndMuscleInfo' 
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    userInfo: 
      type: object
      properties:
        id: 
          type: integer
          format: int64
          example: 13
        username:
          type: string
          example: batiukmaks3000
        email:
          type: string
          example: killer5000@gmail.com
        first_name:
          type: string
          example: Maksym
        last_name:
          type: string
          example: Batiuk
        sex: 
          type: string
          example: male
          enum:
            - male
            - female
        birthday:
          type: string
          format: date
    measurementRecordCreation:
      type: object
      properties:
        name:
          type: string
          example: waist
        record:
          type: number
          format: float
          example: 42.3
    measurementsList:
      type: array
      items:
        $ref: '#/components/schemas/measurementRecordInfo'
    measurementRecordInfo:
      type: object
      properties:
        id: 
          type: integer
          format: int64
          example: 13
        name: 
          type: string
          example: waist
          enum:
            - weight
            - waist 
            - hips
            - chest
        record: 
          type: number
          format: float
          example: 53.2
        date: 
          type: string
          format: date
    userAndMeasurementsInfo:
      allOf:
        - $ref: '#/components/schemas/userInfo'
        - type: object
          properties:
            measurement_records: 
              type: array
              items:
                $ref: "#/components/schemas/graph"
    userAndExerciseInfo:
      allOf:
        - $ref: '#/components/schemas/userInfo'
        - type: object
          properties:
            graph:
              $ref: '#/components/schemas/graph'
            # records:
            #   type: array
            #   items:
            #     $ref: '#/components/schemas/recordInfo'
    userAndMuscleInfo:
      allOf:
        - $ref: '#/components/schemas/userInfo'
        - type: object
          properties:
            graph:
              $ref: '#/components/schemas/graph'
...
    recordCreation:
      type: object
      properties:
        exercise_id:
          type: integer
        sets:
          type: array
          items:
            type: object
            properties:
              sequence_number:
                type: integer
                example: 1
              reps:
                type: integer
                example: 15
              time:
                type: number
                format: float
                example: 0.43
              weight:
                type: number
                format: float
                example: 25
    recordInfo:
      allOf:
        - type: object
          properties:
            id: 
              type: integer
            exercise_name: 
              type: string
        - $ref: '#/components/schemas/recordCreation'
    graph:
      type: object
      properties:
        graph_name:
          type: string
          example: weight
        coordinates:
          type: array
          items:
            type: object
            properties:
              value:
                type: number
                format: float
                example: 12.7
              date:
                type: string
                format: date
                example: 01-03-2004
    error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
  responses:
...
    Unauthorized:
      description: Unauthorized user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...
    NotFound:
      description: The resourse is not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'